requirements.txt
Django==2.1.7
django-tinymce==2.8.0
